Adapt routes correctly and fix ResponseRecorder bug
Tests used gorilla mux but handlers used chi router to get path params.
Because of this difference the handlers weren't able to get the path
params correctly. In gorilla mux the look like this /{bic}/ whereas
chi is using this notations /:bic/.

Other problem was that the ResponseRecorder was used for every single
response. So it contained all the data from all previous responses.
No I'm creating a ResponseRecorder for every single request.